server/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dlib==20.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.120.0
h11==0.16.0
httptools==0.7.1
idna==3.11
mysql-connector-python==9.5.0
numpy==2.3.4
pillow==12.0.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
